+++ src/main/java/com/elibrary/group4/controller/User/UserBorrowController.java
@@ -1,5 +1,7 @@
 package com.elibrary.group4.controller.User;
 
+import com.elibrary.group4.Utils.Validation.JwtUtil;
+import com.elibrary.group4.exception.ForbiddenException;
 import com.elibrary.group4.model.Book;
 import com.elibrary.group4.model.Borrow;
 import com.elibrary.group4.model.User;
@@ -31,6 +33,9 @@ public class UserBorrowController {
     @Autowired
     ModelMapper modelMapper;
 
+    @Autowired
+    JwtUtil jwtUtil;
+
     @PostMapping
     public ResponseEntity create(@Valid @RequestBody BorrowRequest request) throws Exception {
         Borrow borrow = modelMapper.map(request,Borrow.class);
@@ -45,19 +50,38 @@ public class UserBorrowController {
     }
 
     @GetMapping
-    public ResponseEntity get() throws Exception{
-        return ResponseEntity.status(HttpStatus.OK).body(new SuccessResponse<>("Success",borrowService.findAll()));
+    public ResponseEntity get(@RequestHeader("Authorization") String token) throws Exception{
+        var tokenAndRole = jwtUtil.getRoleAndId(token);
+
+        ResponseEntity response = null;
+        if (tokenAndRole.get("role").equals("ADMIN")) {
+            response = ResponseEntity.status(HttpStatus.OK).body(new SuccessResponse<>("Success",borrowService.findAll()));
+        } else if (tokenAndRole.get("role").equals("user")) {
+            response = ResponseEntity.status(HttpStatus.OK).body(new SuccessResponse<>("Success",borrowService.findBorrowByUserId(tokenAndRole.get("userId"))));
+        }
+
+        return response;
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity update(@Valid @RequestBody BorrowRequest request, @PathVariable("id") String id) throws Exception{
+    public ResponseEntity update(@RequestHeader("Authorization") String token, @Valid @RequestBody BorrowRequest request, @PathVariable("id") String id) throws Exception{
+        var tokenAndRole = jwtUtil.getRoleAndId(token);
+
+        if (!tokenAndRole.get("role").equals("ADMIN")) {
+            throw new ForbiddenException("Forbidden");
+        }
         Borrow borrow = modelMapper.map(request, Borrow.class);
         Borrow update = borrowService.update(id,borrow);
         return ResponseEntity.status(HttpStatus.OK).body(new SuccessResponse<>("Updated",update));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity delete(@PathVariable("id")String id) throws Exception{
+    public ResponseEntity delete(@RequestHeader("Authorization") String token, @PathVariable("id")String id) throws Exception{
+        var tokenAndRole = jwtUtil.getRoleAndId(token);
+
+        if (!tokenAndRole.get("role").equals("ADMIN")) {
+            throw new ForbiddenException("Forbidden");
+        }
         borrowService.delete(id);
         return ResponseEntity.status(HttpStatus.OK).body(new SuccessResponse<>("Deleted",null));
     }
